Removed legacy styling - resolves (#174)
