+++ src/components/export/ExportVisualDialogPivot.tsx
@@ -1,11 +1,18 @@
 import * as React from 'react';
-import { Pivot, PivotItem } from '@fluentui/react';
+import { Pivot, PivotItem, IPivotStyles } from '@fluentui/react/lib/Pivot';
+import { IStyleSet } from '@fluentui/merge-styles';
 
 import store from '../../store';
 import { TExportOperation } from '../../core/template';
-import { exportPivotStyles } from '../../config/styles';
 import { i18nValue } from '../../core/ui/i18n';
 import { resolveTemplateExportPivotAria } from '../../core/ui/aria';
+
+const exportPivotStyles: Partial<IStyleSet<IPivotStyles>> = {
+    itemContainer: {
+        marginTop: '10px',
+        height: '100%'
+    }
+};
 
 const ExportVisualDialogPivot = () => {
     const { templateSelectedExportOperation, updateSelectedExportOperation } =
